+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/security/DefaultUserPermissionProvider.kt
@@ -3,33 +3,38 @@ package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.security
 import org.springframework.core.ParameterizedTypeReference
 import org.springframework.web.reactive.function.client.WebClient
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.exception.NoDataAvailableException
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.security.authentication.AuthUser
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.model.Caseload
+import uk.gov.justice.hmpps.kotlin.auth.AuthSource
+
+const val WARNING_NO_ACTIVE_CASELOAD = "User has not set an active caseload."
+const val WARNING_NO_CASELOADS = "User does not have any caseloads."
 
 class DefaultUserPermissionProvider(private val manageUsersWebClient: WebClient) : UserPermissionProvider {
+  override fun getCaseloads(username: String): CaseloadResponse {
+    val caseloadResponse = fetchCaseloadInfo(username)
 
-  override fun getActiveCaseloadId(username: String): String {
-    val caseloadResponse = getPrisonUsersCaseload(username)
+    if (caseloadResponse.caseloads.isEmpty() || caseloadResponse.activeCaseload == null) {
+      val userInfo = getUserInfo(username)
+      if (userInfo.authSource != AuthSource.NOMIS) {
+        return caseloadResponse
+      }
+      if (caseloadResponse.caseloads.isEmpty()) {
+        throw NoDataAvailableException(WARNING_NO_CASELOADS)
+      }
 
-    if (caseloadResponse.accountType != "GENERAL") {
-      throw NoDataAvailableException("'${caseloadResponse.accountType}' account types are currently not supported.")
-    }
-
-    if (caseloadResponse.activeCaseload == null) {
       throw NoDataAvailableException(WARNING_NO_ACTIVE_CASELOAD)
     }
 
-    return caseloadResponse.activeCaseload.id
+    return caseloadResponse
   }
 
-  override fun getCaseloads(username: String): List<Caseload> {
-    val caseloadResponse = getPrisonUsersCaseload(username)
-
-    if (caseloadResponse.caseloads.isEmpty()) {
-      throw NoDataAvailableException(WARNING_NO_CASELOADS)
-    }
-
-    return caseloadResponse.caseloads.sortedBy { it.id }.map { Caseload(it.id, it.name) }
-  }
+  override fun getUserInfo(username: String): AuthUser = manageUsersWebClient.get()
+    .uri("/users/$username")
+    .header("Content-Type", "application/json")
+    .retrieve()
+    .bodyToMono(AuthUser::class.java)
+    .block()!!
 
   override fun getUsersRoles(username: String): List<String> = manageUsersWebClient.get()
     .uri("/users/$username/roles")
@@ -38,11 +43,28 @@ class DefaultUserPermissionProvider(private val manageUsersWebClient: WebClient)
     .bodyToMono(ROLES)
     .block()!!.map { it.roleCode }
 
-  override fun getPrisonUsersCaseload(username: String): CaseloadResponse = manageUsersWebClient.get()
+  fun fetchCaseloadInfo(username: String): CaseloadResponse = manageUsersWebClient.get()
     .uri("/prisonusers/$username/caseloads")
     .header("Content-Type", "application/json")
-    .retrieve()
-    .bodyToMono(CaseloadResponse::class.java)
+    .exchangeToMono { response ->
+      if (response.statusCode().value() == 404) {
+        response.releaseBody().thenReturn(
+          CaseloadResponse(
+            username = username,
+            active = false,
+            accountType = "GENERAL",
+            caseloads = emptyList(),
+            activeCaseload = null,
+          ),
+        )
+      } else {
+        response
+          .bodyToMono(CaseloadResponse::class.java)
+          .map { caseload ->
+            caseload.copy(caseloads = caseload.caseloads.sortedBy { it.id })
+          }
+      }
+    }
     .block()!!
 }
 data class RolesResponse(val roleCode: String)