+++ src/test/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/integration/DataApiIntegrationTest.kt
@@ -1,16 +1,11 @@
 package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.integration
 
-import com.github.tomakehurst.wiremock.client.WireMock
-import com.github.tomakehurst.wiremock.matching.RequestPatternBuilder
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
-import org.springframework.http.HttpHeaders
-import org.springframework.http.HttpStatus
-import org.springframework.http.MediaType
 import org.springframework.test.context.DynamicPropertyRegistry
 import org.springframework.test.context.DynamicPropertySource
 import org.springframework.test.web.reactive.server.StatusAssertions
@@ -37,6 +32,7 @@ import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.ConfiguredApi
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.ExternalMovementEntity
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.PrisonerEntity
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.security.WARNING_NO_ACTIVE_CASELOAD
+import uk.gov.justice.hmpps.kotlin.auth.AuthSource
 import java.time.LocalDateTime
 
 class DataApiIntegrationTest : IntegrationTestBase() {
@@ -95,8 +91,6 @@ class DataApiIntegrationTest : IntegrationTestBase() {
         .isEqualTo(dateTimeWithSeconds(movementPrisoner4[DATE]))
         .jsonPath("$.[1].date")
         .isEqualTo(dateTimeWithSeconds(movementPrisonerDestinationCaseloadDirectionIn[DATE]))
-
-      assertThat(wireMockServer.findAll(RequestPatternBuilder().withUrl("/users/me/caseloads")).size).isEqualTo(2)
     }
 
     @Test
@@ -119,8 +113,6 @@ class DataApiIntegrationTest : IntegrationTestBase() {
         .isEqualTo(dateTimeWithSeconds(movementPrisoner4[DATE]))
         .jsonPath("$.[1].date")
         .isEqualTo(dateTimeWithSeconds(movementPrisonerDestinationCaseloadDirectionIn[DATE]))
-
-      assertThat(wireMockServer.findAll(RequestPatternBuilder().withUrl("/users/me/caseloads")).size).isEqualTo(2)
     }
 
     @Test
@@ -142,8 +134,6 @@ class DataApiIntegrationTest : IntegrationTestBase() {
         .isEqualTo(dateTimeWithSeconds(movementPrisoner4[DATE]))
         .jsonPath("$.[1].date")
         .isEqualTo(dateTimeWithSeconds(movementPrisonerDestinationCaseloadDirectionIn[DATE]))
-
-      assertThat(wireMockServer.findAll(RequestPatternBuilder().withUrl("/users/me/caseloads")).size).isEqualTo(2)
     }
 
     @Test
@@ -167,8 +157,6 @@ class DataApiIntegrationTest : IntegrationTestBase() {
         .isEqualTo(dateTimeWithSeconds(movementPrisoner4[DATE]))
         .jsonPath("$.[1].date")
         .isEqualTo(dateTimeWithSeconds(movementPrisonerDestinationCaseloadDirectionIn[DATE]))
-
-      assertThat(wireMockServer.findAll(RequestPatternBuilder().withUrl("/users/me/caseloads")).size).isEqualTo(2)
     }
   }
 
@@ -197,8 +185,6 @@ class DataApiIntegrationTest : IntegrationTestBase() {
       ]       
       """,
       )
-
-    assertThat(wireMockServer.findAll(RequestPatternBuilder().withUrl("/users/me/caseloads")).size).isEqualTo(2)
   }
 
   @Test
@@ -231,8 +217,6 @@ class DataApiIntegrationTest : IntegrationTestBase() {
         assertThat(body?.startsWith("\uFEFF")).isTrue()
         assertThat(body?.trim()?.replace("\uFEFF", "")).isEqualTo(expected)
       }
-
-    assertThat(wireMockServer.findAll(RequestPatternBuilder().withUrl("/users/me/caseloads")).size).isEqualTo(2)
   }
 
   class ConfiguredApiFormulaTest : IntegrationTestBase() {
@@ -272,8 +256,6 @@ class DataApiIntegrationTest : IntegrationTestBase() {
       ]       
       """,
         )
-
-      assertThat(wireMockServer.findAll(RequestPatternBuilder().withUrl("/users/me/caseloads")).size).isEqualTo(2)
     }
 
     @Test
@@ -318,8 +300,6 @@ class DataApiIntegrationTest : IntegrationTestBase() {
       ]       
       """,
         )
-
-      assertThat(wireMockServer.findAll(RequestPatternBuilder().withUrl("/users/me/caseloads")).size).isEqualTo(2)
     }
   }
 
@@ -442,14 +422,8 @@ class DataApiIntegrationTest : IntegrationTestBase() {
   @Test
   fun `Data API returns empty list and warning header if no active caseloads`() {
     wireMockServer.resetAll()
-    wireMockServer.stubFor(
-      WireMock.get("/users/me/caseloads").willReturn(
-        WireMock.aResponse()
-          .withStatus(HttpStatus.OK.value())
-          .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-          .withBody(caseloadsWithNoneActive),
-      ),
-    )
+    manageUsersMockServer.stubLookupUserCaseload("request-user", null)
+    manageUsersMockServer.stubGetUserInfo("request-user")
     stubDefinitionsResponse()
 
     webTestClient.get()
@@ -473,16 +447,36 @@ class DataApiIntegrationTest : IntegrationTestBase() {
   }
 
   @Test
-  fun `Data API count returns zero and warning header if no active caseloads`() {
+  fun `Data API returns empty list and no warning header if no active caseloads and not NOMIS user`() {
     wireMockServer.resetAll()
-    wireMockServer.stubFor(
-      WireMock.get("/users/me/caseloads").willReturn(
-        WireMock.aResponse()
-          .withStatus(HttpStatus.OK.value())
-          .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-          .withBody(caseloadsWithNoneActive),
-      ),
-    )
+    manageUsersMockServer.stubLookupUserCaseload("request-user", null)
+    manageUsersMockServer.stubGetUserInfo("request-user", AuthSource.DELIUS)
+    stubDefinitionsResponse()
+
+    webTestClient.get()
+      .uri { uriBuilder: UriBuilder ->
+        uriBuilder
+          .path("/reports/external-movements/last-month")
+          .queryParam("selectedPage", 1)
+          .queryParam("pageSize", 3)
+          .queryParam("sortColumn", "date")
+          .queryParam("sortedAsc", false)
+          .build()
+      }
+      .headers(setAuthorisation(roles = listOf(authorisedRole)))
+      .exchange()
+      .expectStatus()
+      .isOk()
+      .expectHeader()
+      .doesNotExist(ResponseHeader.NO_DATA_WARNING_HEADER_NAME)
+      .expectBody()
+      .json("[]")
+  }
+
+  @Test
+  fun `Data API count returns zero and warning header if no active caseloads`() {
+    manageUsersMockServer.stubLookupUserCaseload("request-user", null)
+    manageUsersMockServer.stubGetUserInfo("request-user")
     stubDefinitionsResponse()
 
     webTestClient.get()