+++ src/test/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/integration/IntegrationSystemTestBase.kt
@@ -29,9 +29,11 @@ import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.establishment
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.integration.IntegrationTestBase.Companion.TEST_TOKEN
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.integration.wiremock.HmppsAuthMockServer
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.integration.wiremock.ManageUsersMockServer
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.security.CaseloadResponse
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.security.DprSystemAuthAwareAuthenticationToken
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.AsyncDataApiService
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.model.Caseload
+import uk.gov.justice.hmpps.kotlin.auth.AuthSource
 import uk.gov.justice.hmpps.test.kotlin.auth.JwtAuthorisationHelper
 
 @SpringBootTest(webEnvironment = RANDOM_PORT, properties = ["spring.main.allow-bean-definition-overriding=true"])
@@ -90,10 +92,7 @@ abstract class IntegrationSystemTestBase {
     @JvmStatic
     fun startMocks() {
       hmppsAuthMockServer.start()
-      hmppsAuthMockServer.stubGrantToken()
       manageUsersMockServer.start()
-
-      hmppsAuthMockServer.stubGrantToken()
     }
 
     @AfterAll
@@ -122,11 +121,44 @@ abstract class IntegrationSystemTestBase {
     whenever(jwt.tokenValue).then { TEST_TOKEN }
     whenever(authentication.jwt).then { jwt }
     authenticationHelper.authentication = authentication
+    whenever(authentication.authSource).then { AuthSource.NOMIS }
     whenever(authentication.getRoles()).thenReturn(listOf(authorisedRole))
-    whenever(authentication.getCaseLoads()).thenReturn(listOf(Caseload("WWI", "WANDSWORTH (HMP)")))
+    whenever(authentication.getCaseLoads()).thenReturn(
+      CaseloadResponse(
+        username = "request-user",
+        active = true,
+        accountType = "GENERAL",
+        activeCaseload = Caseload(id = "WWI", name = "WANDSWORTH (HMP)"),
+        caseloads = listOf(Caseload("WWI", "WANDSWORTH (HMP)")),
+      ),
+    )
     whenever(authentication.getActiveCaseLoadId()).thenReturn("WWI")
     whenever(authentication.getCaseLoadIds()).thenReturn(listOf("WWI"))
     whenever(authentication.getUsername()).thenReturn("request-user")
+
+    hmppsAuthMockServer.stubGrantToken()
+    manageUsersMockServer.stubLookupUsersRoles("request-user", listOf("INCIDENT_REPORTS__RO", "PRISONS_REPORTING_USER"))
+    manageUsersMockServer.stubLookupUserCaseload(
+      "request-user",
+      "ABC",
+      """
+        "caseloads": [
+          {
+            "id": "ABC",
+            "name": "ABCPRISON (ABC)"
+          },
+          {
+            "id": "DEF",
+            "name": "DEFPRISON (DEF)"
+          },
+          {
+            "id": "GHI",
+            "name": "GHIPRISON (GHI)"
+          }
+        ]
+      """.trimIndent(),
+    )
+    manageUsersMockServer.stubGetUserInfo("request-user")
   }
 
   protected fun setAuthorisation(